refactor: API 서비스 구조 변경
- 잘못된 서비스 경로를 '/services'에서 '/lib/api'로 변경
- 관련 훅에서 API 경로 업데이트
