+++ lib/api/talentDetail.ts
